BiasedDoom 4.15.8: huge-map coordinates through +/-262144, UDMF sector
heights beyond int16, fog and visibility scaling, huge-map sky
rendering, and procedural maps up to size 160.

+++ src/maploader/renderinfo.cpp
@@ -53,7 +53,9 @@ struct MapSectionGenerator
 	{
 		double X, Y;
 
-		operator int() const { return xs_FloorToInt(X) + 65536 * xs_FloorToInt(Y); }
+		// Hash for the TMap key. Computed in 64 bit so that coordinates beyond
+		// the old +/-32768 range cannot overflow into undefined behavior.
+		operator int() const { return (int)(int64_t(xs_FloorToInt(X)) + 65536LL * int64_t(xs_FloorToInt(Y))); }
 		bool operator!= (const cvertex_t &other) const { return fabs(X - other.X) >= EQUAL_EPSILON || fabs(Y - other.Y) >= EQUAL_EPSILON; }
 		cvertex_t& operator =(const vertex_t *v) { X = v->fX(); Y = v->fY(); return *this; }
 	};
